refactor(bootstrap): A66 step 5 - extract phases 0-4 into phases_assembly.go
将 phase0ValidateOptions / validateHealthCheckers / phase1LoadConfig /
phase2InitPubSub / samePubSubIdentity / phase3InitAssembly /
phase4WireAuthAndWatcher / bindConfigWatcher / buildOnChangeCallback /
applyConfigReload / syncObservedGeneration / notifyCellsConfigChanged /
shouldNotifyCell / buildCellConfigSnap / invokeCellReload 这一组
config-load → assembly-init → auth-wire 的 phase 方法从 bootstrap_phases.go
搬到独立文件 phases_assembly.go (388 行)，清理 bootstrap_phases.go 中的
reflect / eventbus 残留 import。bootstrap_phases.go 保留 phaseState
类型定义 + phase5+ HTTP/router/event/worker/shutdown 阶段。

同步重新生成 assemblies/corebundle/generated/metrics-schema.yaml：
http_request_duration_seconds / http_requests_total 在
bootstrap_phases.go 的行号从 991 降至 621（搬出 phase 0-4 后）。

ref: uber-go/fx app.go — sequential startup with transactional rollback
ref: go-micro/v2 config — OnChange callback binding after watcher is created
